+++ react/components/GalleryRow.tsx
@@ -7,6 +7,7 @@ import { useRenderOnView } from '../hooks/useRenderOnView'
 import GalleryItem from './GalleryItem'
 import type { Product } from '../Gallery'
 import type { MobileLayoutMode } from '../GalleryLegacy'
+import type { PreferredSKU } from '../GalleryLayout'
 
 const CSS_HANDLES = ['galleryItem'] as const
 
@@ -24,6 +25,8 @@ interface GalleryRowProps {
   listName: string
   customSummaryInterval?: number
   CustomSummary?: ComponentType
+  /** Logic to enable which SKU will be the selected item */
+  preferredSKU?: PreferredSKU
 }
 
 function GalleryRow({
@@ -36,6 +39,7 @@ function GalleryRow({
   listName,
   customSummaryInterval,
   CustomSummary,
+  preferredSKU
 }: GalleryRowProps) {
   const handles = useCssHandles(CSS_HANDLES)
 
@@ -66,7 +70,7 @@ function GalleryRow({
 
         return (
           <div
-            key={product.productId}
+            key={product.cacheId}
             style={style}
             className={classNames(
               applyModifiers(handles.galleryItem, [
@@ -83,6 +87,7 @@ function GalleryRow({
               position={absoluteProductIndex}
               listName={listName}
               CustomSummary={shouldRenderCustom ? CustomSummary : undefined}
+              preferredSKU={preferredSKU}
             />
           </div>
         )